Refactor Item out to own file
