+++ todo_app/data/trello_items.py
@@ -1,4 +1,4 @@
-from enum import Enum
+from .item import Status, Item
 import requests
 import os
 
@@ -8,19 +8,6 @@ BASE_PAYLOAD = {
     'key': os.getenv('API_KEY'),
     'token': os.getenv('TOKEN')
 }
-class Status(Enum):
-    TO_DO = "To Do"
-    DONE = "Done"
-
-class Item:
-    def __init__(self, id, title, status = Status.TO_DO.value):
-        self.id = id
-        self.title = title
-        self.status = status
-    
-    @classmethod
-    def from_trello_card(cls, card, list):
-        return cls(card['id'], card['name'], list['name'])
 
 def get_items():
     payload = {**BASE_PAYLOAD, 'cards': 'open'}
